rename pkg, fix rbac, len(items) = 0
