+++ pkg/kubedb/mongodb.go
@@ -1,10 +1,11 @@
-package db_broker
+package kubedb
 
 import (
 	"fmt"
 	"strings"
 
 	jsonTypes "github.com/appscode/go/encoding/json/types"
+	"github.com/appscode/go/types"
 	"github.com/golang/glog"
 	api "github.com/kubedb/apimachinery/apis/kubedb/v1alpha1"
 	cs "github.com/kubedb/apimachinery/client/clientset/versioned/typed/kubedb/v1alpha1"
@@ -16,64 +17,76 @@ import (
 	appcat "kmodules.xyz/custom-resources/apis/appcatalog/v1alpha1"
 )
 
-type MySQLProvider struct {
+type MongoDbProvider struct {
 	extClient        cs.KubedbV1alpha1Interface
 	storageClassName string
 }
 
-func NewMySQLProvider(config *rest.Config, storageClassName string) Provider {
-	return &MySQLProvider{
+func NewMongoDbProvider(config *rest.Config, storageClassName string) Provider {
+	return &MongoDbProvider{
 		extClient:        cs.NewForConfigOrDie(config),
 		storageClassName: storageClassName,
 	}
 }
 
-func demoMySQLSpec() api.MySQLSpec {
-	return api.MySQLSpec{
-		Version:           jsonTypes.StrYo(demoMySQLVersion),
+func demoMongoDBSpec() api.MongoDBSpec {
+	return api.MongoDBSpec{
+		Version:           jsonTypes.StrYo(demoMongoDBVersion),
 		StorageType:       api.StorageTypeEphemeral,
 		TerminationPolicy: api.TerminationPolicyWipeOut,
 	}
 }
 
-func (p MySQLProvider) Metadata() (string, string) {
-	return "kubedb", "mysql"
+func demoMongoDBClusterSpec() api.MongoDBSpec {
+	mgSpec := demoMongoDBSpec()
+	mgSpec.Replicas = types.Int32P(3)
+	mgSpec.ReplicaSet = &api.MongoDBReplicaSet{
+		Name: "rs0",
+	}
+
+	return mgSpec
 }
 
-func (p MySQLProvider) Create(provisionInfo ProvisionInfo) error {
-	var my api.MySQL
+func (p MongoDbProvider) Metadata() (string, string) {
+	return "kubedb", "mongodb"
+}
+
+func (p MongoDbProvider) Create(provisionInfo ProvisionInfo) error {
+	var mg api.MongoDB
 
 	// set metadata from provision info
-	if err := provisionInfo.applyToMetadata(&my.ObjectMeta); err != nil {
+	if err := provisionInfo.applyToMetadata(&mg.ObjectMeta); err != nil {
 		return err
 	}
 
 	// set postgres spec
 	switch provisionInfo.PlanID {
-	case planMySQLDemo:
-		my.Spec = demoMySQLSpec()
-	case planMySQL:
-		if err := provisionInfo.applyToSpec(&my.Spec); err != nil {
+	case planMongoDBDemo:
+		mg.Spec = demoMongoDBSpec()
+	case planMongoDBClusterDemo:
+		mg.Spec = demoMongoDBClusterSpec()
+	case planMongoDB:
+		if err := provisionInfo.applyToSpec(&mg.Spec); err != nil {
 			return err
 		}
 	}
 
-	glog.Infof("Creating mysql obj %q in namespace %q...", my.Name, my.Namespace)
-	_, err := p.extClient.MySQLs(my.Namespace).Create(&my)
+	glog.Infof("Creating mongodb obj %q in namespace %q...", mg.Name, mg.Namespace)
+	_, err := p.extClient.MongoDBs(mg.Namespace).Create(&mg)
 
 	return err
 }
 
-func (p MySQLProvider) Delete(name, namespace string) error {
-	glog.Infof("Deleting mysql obj %q from namespace %q...", name, namespace)
+func (p MongoDbProvider) Delete(name, namespace string) error {
+	glog.Infof("Deleting mongodb obj %q from namespace %q...", name, namespace)
 
-	my, err := p.extClient.MySQLs(namespace).Get(name, metav1.GetOptions{})
+	mg, err := p.extClient.MongoDBs(namespace).Get(name, metav1.GetOptions{})
 	if err != nil {
 		return err
 	}
 
-	if my.Spec.TerminationPolicy != api.TerminationPolicyWipeOut {
-		if err := patchMySQL(p.extClient, my, func(in *api.MySQL) *api.MySQL {
+	if mg.Spec.TerminationPolicy != api.TerminationPolicyWipeOut {
+		if err := patchMongoDb(p.extClient, mg, func(in *api.MongoDB) *api.MongoDB {
 			in.Spec.TerminationPolicy = api.TerminationPolicyWipeOut
 			return in
 		}); err != nil {
@@ -81,14 +94,14 @@ func (p MySQLProvider) Delete(name, namespace string) error {
 		}
 	}
 
-	if err := p.extClient.MySQLs(namespace).Delete(name, &metav1.DeleteOptions{}); err != nil {
+	if err := p.extClient.MongoDBs(namespace).Delete(name, &metav1.DeleteOptions{}); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-func (p MySQLProvider) Bind(
+func (p MongoDbProvider) Bind(
 	app *appcat.AppBinding,
 	params map[string]interface{},
 	data map[string]interface{}) (*Credentials, error) {
@@ -128,24 +141,24 @@ func (p MySQLProvider) Bind(
 	}, nil
 }
 
-func (p MySQLProvider) GetProvisionInfo(instanceID string) (*ProvisionInfo, error) {
-	mysqls, err := p.extClient.MySQLs(corev1.NamespaceAll).List(metav1.ListOptions{
+func (p MongoDbProvider) GetProvisionInfo(instanceID string) (*ProvisionInfo, error) {
+	mongodbs, err := p.extClient.MongoDBs(corev1.NamespaceAll).List(metav1.ListOptions{
 		LabelSelector: labels.Set{
 			InstanceKey: instanceID,
 		}.String(),
 	})
-	if err != nil {
+	if err != nil || len(mongodbs.Items) == 0 {
 		return nil, err
 	}
 
-	if len(mysqls.Items) > 1 {
+	if len(mongodbs.Items) > 1 {
 		var instances []string
-		for _, mysql := range mysqls.Items {
-			instances = append(instances, fmt.Sprintf("%s/%s", mysql.Namespace, mysql.Name))
+		for _, mongodb := range mongodbs.Items {
+			instances = append(instances, fmt.Sprintf("%s/%s", mongodb.Namespace, mongodb.Name))
 		}
 
-		return nil, errors.Errorf("%d MySQLs with instance id %s found: %s",
-			len(mysqls.Items), instanceID, strings.Join(instances, ", "))
+		return nil, errors.Errorf("%d MongoDBs with instance id %s found: %s",
+			len(mongodbs.Items), instanceID, strings.Join(instances, ", "))
 	}
-	return provisionInfoFromObjectMeta(mysqls.Items[0].ObjectMeta)
+	return provisionInfoFromObjectMeta(mongodbs.Items[0].ObjectMeta)
 }
