+++ pkg/kubedb/mysql.go
@@ -1,90 +1,79 @@
-package db_broker
+package kubedb
 
 import (
 	"fmt"
 	"strings"
 
 	jsonTypes "github.com/appscode/go/encoding/json/types"
-	"github.com/appscode/go/types"
 	"github.com/golang/glog"
 	api "github.com/kubedb/apimachinery/apis/kubedb/v1alpha1"
 	cs "github.com/kubedb/apimachinery/client/clientset/versioned/typed/kubedb/v1alpha1"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	k8sLabels "k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/rest"
 	appcat "kmodules.xyz/custom-resources/apis/appcatalog/v1alpha1"
 )
 
-type PostgreSQLProvider struct {
+type MySQLProvider struct {
 	extClient        cs.KubedbV1alpha1Interface
 	storageClassName string
 }
 
-func NewPostgreSQLProvider(config *rest.Config, storageClassName string) Provider {
-	return &PostgreSQLProvider{
+func NewMySQLProvider(config *rest.Config, storageClassName string) Provider {
+	return &MySQLProvider{
 		extClient:        cs.NewForConfigOrDie(config),
 		storageClassName: storageClassName,
 	}
 }
 
-func demoPostgresSpec() api.PostgresSpec {
-	return api.PostgresSpec{
-		Version:           jsonTypes.StrYo(demoPostgresVersion),
-		Replicas:          types.Int32P(1),
+func demoMySQLSpec() api.MySQLSpec {
+	return api.MySQLSpec{
+		Version:           jsonTypes.StrYo(demoMySQLVersion),
 		StorageType:       api.StorageTypeEphemeral,
 		TerminationPolicy: api.TerminationPolicyWipeOut,
 	}
 }
 
-func demoHAPostgresSpec() api.PostgresSpec {
-	pgSpec := demoPostgresSpec()
-	pgSpec.Replicas = types.Int32P(3)
-
-	return pgSpec
+func (p MySQLProvider) Metadata() (string, string) {
+	return "kubedb", "mysql"
 }
 
-func (p PostgreSQLProvider) Metadata() (string, string) {
-	return "kubedb", "postgresql"
-}
-
-func (p PostgreSQLProvider) Create(provisionInfo ProvisionInfo) error {
-	var pg api.Postgres
+func (p MySQLProvider) Create(provisionInfo ProvisionInfo) error {
+	var my api.MySQL
 
 	// set metadata from provision info
-	if err := provisionInfo.applyToMetadata(&pg.ObjectMeta); err != nil {
+	if err := provisionInfo.applyToMetadata(&my.ObjectMeta); err != nil {
 		return err
 	}
 
 	// set postgres spec
 	switch provisionInfo.PlanID {
-	case planPostgresDemo:
-		pg.Spec = demoPostgresSpec()
-	case planPostgresHADemo:
-		pg.Spec = demoHAPostgresSpec()
-	case planPostgres:
-		if err := provisionInfo.applyToSpec(&pg.Spec); err != nil {
+	case planMySQLDemo:
+		my.Spec = demoMySQLSpec()
+	case planMySQL:
+		if err := provisionInfo.applyToSpec(&my.Spec); err != nil {
 			return err
 		}
 	}
 
-	glog.Infof("Creating postgres obj %q in namespace %q...", pg.Name, pg.Namespace)
-	_, err := p.extClient.Postgreses(pg.Namespace).Create(&pg)
+	glog.Infof("Creating mysql obj %q in namespace %q...", my.Name, my.Namespace)
+	_, err := p.extClient.MySQLs(my.Namespace).Create(&my)
 
 	return err
 }
 
-func (p PostgreSQLProvider) Delete(name, namespace string) error {
-	glog.Infof("Deleting postgres obj %q from namespace %q...", name, namespace)
+func (p MySQLProvider) Delete(name, namespace string) error {
+	glog.Infof("Deleting mysql obj %q from namespace %q...", name, namespace)
 
-	pgsql, err := p.extClient.Postgreses(namespace).Get(name, metav1.GetOptions{})
+	my, err := p.extClient.MySQLs(namespace).Get(name, metav1.GetOptions{})
 	if err != nil {
 		return err
 	}
 
-	if pgsql.Spec.TerminationPolicy != api.TerminationPolicyWipeOut {
-		if err := patchPostgreSQL(p.extClient, pgsql, func(in *api.Postgres) *api.Postgres {
+	if my.Spec.TerminationPolicy != api.TerminationPolicyWipeOut {
+		if err := patchMySQL(p.extClient, my, func(in *api.MySQL) *api.MySQL {
 			in.Spec.TerminationPolicy = api.TerminationPolicyWipeOut
 			return in
 		}); err != nil {
@@ -92,14 +81,14 @@ func (p PostgreSQLProvider) Delete(name, namespace string) error {
 		}
 	}
 
-	if err := p.extClient.Postgreses(namespace).Delete(name, &metav1.DeleteOptions{}); err != nil {
+	if err := p.extClient.MySQLs(namespace).Delete(name, &metav1.DeleteOptions{}); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-func (p PostgreSQLProvider) Bind(
+func (p MySQLProvider) Bind(
 	app *appcat.AppBinding,
 	params map[string]interface{},
 	data map[string]interface{}) (*Credentials, error) {
@@ -139,24 +128,24 @@ func (p PostgreSQLProvider) Bind(
 	}, nil
 }
 
-func (p PostgreSQLProvider) GetProvisionInfo(instanceID string) (*ProvisionInfo, error) {
-	postgreses, err := p.extClient.Postgreses(corev1.NamespaceAll).List(metav1.ListOptions{
-		LabelSelector: k8sLabels.Set{
+func (p MySQLProvider) GetProvisionInfo(instanceID string) (*ProvisionInfo, error) {
+	mysqls, err := p.extClient.MySQLs(corev1.NamespaceAll).List(metav1.ListOptions{
+		LabelSelector: labels.Set{
 			InstanceKey: instanceID,
 		}.String(),
 	})
-	if err != nil {
+	if err != nil || len(mysqls.Items) == 0 {
 		return nil, err
 	}
 
-	if len(postgreses.Items) > 1 {
+	if len(mysqls.Items) > 1 {
 		var instances []string
-		for _, postgres := range postgreses.Items {
-			instances = append(instances, fmt.Sprintf("%s/%s", postgres.Namespace, postgres.Name))
+		for _, mysql := range mysqls.Items {
+			instances = append(instances, fmt.Sprintf("%s/%s", mysql.Namespace, mysql.Name))
 		}
 
-		return nil, errors.Errorf("%d Postgreses with instance id %s found: %s",
-			len(postgreses.Items), instanceID, strings.Join(instances, ", "))
+		return nil, errors.Errorf("%d MySQLs with instance id %s found: %s",
+			len(mysqls.Items), instanceID, strings.Join(instances, ", "))
 	}
-	return provisionInfoFromObjectMeta(postgreses.Items[0].ObjectMeta)
+	return provisionInfoFromObjectMeta(mysqls.Items[0].ObjectMeta)
 }
