+++ pkg/kubedb/provider.go
@@ -1,4 +1,4 @@
-package db_broker
+package kubedb
 
 import (
 	"encoding/json"
